+++ changelog.md
@@ -27,3 +27,4 @@
 - added /supplementaries registry command, lists all the content of a registry
 - improved turn table and wrench rotation logic fixing some bugs and inconsistencies
 - Fixed some dupes with falling urns
+- added cannon map marker